% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nof1.summary.R
\name{kernel_plot}
\alias{kernel_plot}
\title{Kernel density estimation plot}
\usage{
kernel_plot(result, comp = T, ...)
}
\arguments{
\item{result}{Modeling result of class \code{nof1.result} produced by \code{nof1.run}.}

\item{...}{parameters to pass to \code{geom_histogram}.}
}
\description{
Creates a kernel density estimation plot for a specific outcome
}
