% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nof1.R
\docType{package}
\name{nof1-package}
\alias{nof1-package}
\title{mcnet: A package for N of 1 study analysis using Bayesian methods}
\description{
A package for running N of 1 study trials
}
\details{
An N of 1 trial is a clinical trial in which a single patient is the entire trial, a single case study.
The main purpose of this package was to serve as an analysis tool for one of the PCORI grants we were working with.
It is designed for N of 1 trials and can fit bayesian versions of linear regression, logistic/ordinal regression, and poisson regression.
Package includes number of different plotting tools for visualization.
}
