% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nof1.data.R
\name{nof1.data}
\alias{nof1.data}
\title{Make an N of 1 object containing data, priors, and a jags model file}
\usage{
nof1.data(Y, Treat, response = NULL, ncat = NULL, bs.trend = F,
  y.time = NULL, knots.bt.block = NULL, block.no = NULL,
  alpha.prior = NULL, beta.prior = NULL, eta.prior = NULL,
  dc.prior = NULL, c1.prior = NULL, rho.prior = NULL,
  hy.prior = NULL)
}
\arguments{
\item{Y}{Outcome of the study. This should be a vector with length of total number of observations.}

\item{Treat}{Treatment indicator vector with same length as the outcome.}

\item{response}{Type of outcome. Can be normal, binomial, poisson or ordinal.}

\item{ncat}{Number of categories. Used in ordinal models.}

\item{y.time}{parameter used for modelling splines. Time on the original scale with same length as the outcome. Still under development.}

\item{alpha.prior}{Prior for the intercept of the model.}

\item{beta.prior}{Prior for the treatment coefficient.}

\item{eta.prior}{Prior for modelling splines. Still under development.}

\item{dc.prior}{Prior for the length between cutpoints. Used only for ordinal logistic models.}

\item{c1.prior}{Prior for the first cutpoint. Used only for ordinal logistic models.}

\item{rho.prior}{Prior for the correlated error model. Still under development.}

\item{hy.prior}{Prior for the heterogeneity parameter. Supports uniform, gamma, and half normal for normal and binomial response and wishart for multinomial response. It should be a list of length 3, where first element should be the distribution (one of dunif, dgamma, dhnorm, dwish) and the next two are the parameters associated with the distribution. For example, list("dunif", 0, 5) give uniform prior with lower bound 0 and upper bound 5 for the heterogeneity parameter. For wishart distribution, the last two parameter would be the scale matrix and the degrees of freedom.}

\item{knots}{parameter used for modelling splines. Still under development.}
}
\value{
Creates list of variables that are used to run the model using \code{\link{nof1.run}}
\item{Y}{Outcome}
\item{Treat}{Treatment}
\item{baseline}{Baseline variable}
\item{ncat}{Number of categories for ordinal response}
\item{nobs}{Total number of observations in a study}
\item{Treat.name}{Treatment name besides baseline treatment}
\item{response}{The type of response variable}
\item{priors}{Priors that the code will be using. Default priors are used if prior was not specified}
\item{code}{Rjags model file code that is generated using information provided by the user. To view model file inside R, use \code{cat(nof1$code).}}
}
\description{
Make an N of 1 object containing data, priors, and a jags model file
}
\examples{
###Blocker data example
laughter
Y <- laughter$Y
Treat <- laughter$Treat
nof1 <- nof1.data(Y, Treat, ncat = 11, response = "ordinal")
str(nof1)
cat(nof1$code)
}
