% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nof1.data.R
\name{nof1.nma.data}
\alias{nof1.nma.data}
\title{Make an N of 1 object containing data, priors, and a jags model file for (network) meta-analyses}
\usage{
nof1.nma.data(
  Y,
  Treat,
  baseline.Treat,
  ID,
  response,
  model.linkfunc = NULL,
  model.intcpt = "fixed",
  model.slp = "random",
  ord.ncat = NULL,
  ord.model,
  ord.parallel = NULL,
  strata.cov = NULL,
  adjust.strata.cov = NULL,
  lvl2.cov = NULL,
  spline.trend = F,
  trend.type,
  y.time = NULL,
  knots = NULL,
  trend.df = NULL,
  step.trend = F,
  y.step = NULL,
  corr.y = F,
  alpha.prior = NULL,
  beta.prior = NULL,
  eta.prior = NULL,
  dc.prior = NULL,
  c1.prior = NULL,
  rho.prior = NULL,
  hy.prior = NULL,
  na.rm = T,
  ...
)
}
\arguments{
\item{Y}{Outcome of the study. This should be a vector with \code{NA}'s included in time order.}

\item{Treat}{Treatment indicator vector with the same length as the outcome. Can be character or numeric.}

\item{baseline.Treat}{Name of the reference treatment.}

\item{ID}{Participant ID vector with the same length as the outcome.}

\item{response}{Type of outcome. Can be "normal" for continuous outcome, "binomial" for binary outcome,
"poisson" for count outcome, or "ordinal" for ordinal or nominal outcome.}

\item{model.linkfunc}{Link function in the model.}

\item{model.intcpt}{Form of intercept.}

\item{model.slp}{Form of slope. For link function and the forms of intercept and slopes, currently implemented for
1) normal response: "identity"-"fixed/random"-"random", 2) poisson response: "log"-"fixed"-"random",
3) binomial response: "log/logit"-"fixed"-"random".}

\item{ord.ncat}{Number of categories in ordinal response. The parameters for ordinal outcomes need to be tested.}

\item{ord.model}{Used for ordinal outcome to pick the model. Can be "cumulative" for cumulative probability model
or "acat" for adjacent category model.}

\item{ord.parallel}{Whether or not the comparisons between categories or cumulative probabilities are parallel.}

\item{strata.cov}{Only applicable when random intercept model is used. Stratification covariates used during randomization.
Should be a data frame with the columns being the covariates and the number of rows should be equal to the length of the
outcome. Must be of factor type if categorical covariates.}

\item{adjust.strata.cov}{Only applicable when random intercept model is used. A vector with each element taking on possible
values from \code{c("fixed", "random")} indicating how the stratification covariate should be adjusted in the model.}

\item{lvl2.cov}{Participant level covariates for heterogeneous treatment effects.
Should be a data frame with the columns being the covariates and the number of rows should be equal to the length of the outcome.
For fixed-intercept model, participant level covariates will have interaction with treatment (slope) because
there are fixed-intercepts adjusting for each participant in the model; for random-intercept model, participant level covariates
 will have interaction with all treatment-specific intercepts.}

\item{spline.trend}{Indicator for whether the model should adjust for trend using splines. The default
is \code{F}.}

\item{trend.type}{"bs" for basis splines or "ns" for natural splines.}

\item{y.time}{Time when the outcome is measured. Required when adjusting for trend or correlation
(\code{spline.trend}/\code{step.trend}/\code{corr.y} is \code{TRUE}). Should be a vector of the same length as the outcome.}

\item{knots}{Knots in \code{y.time} when adjusting for trend using splines. For \code{trend.type = "bs"}, knots should be set
at the end of each block except for the last block; for or \code{trend.type = "ns"}, knots should be set at the end of each
period except for the last period.}

\item{trend.df}{Degrees of freedom for modeling splines when knots are not set.}

\item{step.trend}{Indicator for whether to adjust for trend using step functions for each period.}

\item{y.step}{Period number corresponding to the outcome. Should be a vector of the same length of the outcome.}

\item{corr.y}{Indicator for whether the correlation among measurements shoule be modeled. The default is
\code{F}.}

\item{alpha.prior}{Prior for the fixed-intercepts.}

\item{beta.prior}{Prior for random intercepts or slopes.}

\item{eta.prior}{Prior for modelling spline terms or heterogeneous treatment effects.}

\item{dc.prior}{Prior for the length between cutpoints. Used only for ordinal logistic models.}

\item{c1.prior}{Prior for the first cutpoint. Used only for ordinal logistic models.}

\item{rho.prior}{Prior for the correlation between random effects or the correlation among repeated measurements on each participant.}

\item{hy.prior}{Prior for the variance of the residual errors for normal response,
the standard deviation of random slopes for binary outcome, the variance of random effects for other types of outcomes.}
}
\description{
Make an N of 1 object containing data, priors, and a jags model file for (network) meta-analyses
}
