% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nof1.simulation.R
\name{nof1.normal.simulation}
\alias{nof1.normal.simulation}
\title{Normal data simulation}
\usage{
nof1.normal.simulation(Base.size = 2, Treat.size = 8, prec = 0.5,
  alpha = 50, beta_A = -3, beta_B = -1)
}
\description{
Simulating sample normal data
}
