% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nof1.summary.R
\name{probability_barplot}
\alias{probability_barplot}
\title{Posterior probability barplot}
\usage{
probability_barplot(result.list)
}
\arguments{
\item{result.list}{A list of one or more modeling results from \code{nof1.run}.}
}
\description{
Creates a posterior probability barplot for treatments
}
