% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nof1.data.R
\name{nof1.data}
\alias{nof1.data}
\title{Make an N of 1 object containing data, priors, and a jags model file}
\usage{
nof1.data(
  Y,
  Treat,
  ord.baseline.Treat = NULL,
  ord.model = NULL,
  response = NULL,
  ncat = NULL,
  bs.trend = F,
  y.time = NULL,
  knots.bt.block = NULL,
  block.no = NULL,
  bs.df = NULL,
  corr.y = F,
  alpha.prior = NULL,
  beta.prior = NULL,
  eta.prior = NULL,
  dc.prior = NULL,
  c1.prior = NULL,
  rho.prior = NULL,
  hy.prior = NULL,
  ...
)
}
\arguments{
\item{Y}{Outcome of the study. This should be a vector with \code{NA}'s included in time order.}

\item{Treat}{Treatment indicator vector with same length as the outcome. Can be character or numeric.}

\item{ord.baseline.Treat}{Used for ordinal outcome to define a reference treatment level.}

\item{ord.model}{Used for ordinal outcome to pick the model. Can be "cumulative" for proportional odds model
or "acat" for restricted adjacent category model.}

\item{response}{Type of outcome. Can be "normal" for continuous outcome, "binomial" for binary outcome,
"poisson" for count outcome, or "ordinal" for ordinal or nominal outcome.}

\item{ncat}{Number of categories. Used in ordinal models.}

\item{bs.trend}{Indicator for whether the model should adjust for trend using splines. The default
is \code{F}.}

\item{y.time}{Parameter used for modeling splines. Time when the outcome is measured.}

\item{knots.bt.block}{parameter used for modeling splines. Indicator for whether or not knots should be set
at the end of each block except for the last block. If \code{TRUE}, user should then specify \code{block.no};
if \code{FALSE}, user should then specify \code{bs.df}.}

\item{block.no}{Block number used for modeling splines for the setting where the knots are set at the end
of each block. Block number with the same length as the outcome.}

\item{bs.df}{Degrees of freedom for modeling splines when knots are not set at the end of each block.}

\item{corr.y}{Indicator for whether the correlation among measurements shoule be modeled. The default is
\code{F}.}

\item{alpha.prior}{Prior for the intercept of the model. Not needed now since we are using treatment-specific
intercept model.}

\item{beta.prior}{Prior for the treatment-specific intercept.}

\item{eta.prior}{Prior for modelling spline terms.}

\item{dc.prior}{Prior for the length between cutpoints. Used only for ordinal logistic models.}

\item{c1.prior}{Prior for the first cutpoint. Used only for ordinal logistic models.}

\item{rho.prior}{Prior for the correlated error model.}

\item{hy.prior}{Prior for the heterogeneity parameter. Supports uniform, gamma, and half normal for
normal and binomial response and wishart for multinomial response. It should be a list of length 3,
where first element should be the distribution (one of dunif, dgamma, dhnorm, dwish) and the next
two are the parameters associated with the distribution. For example, list("dunif", 0, 5) give
uniform prior with lower bound 0 and upper bound 5 for the heterogeneity parameter. For wishart
distribution, the last two parameter would be the scale matrix and the degrees of freedom.}

\item{...}{Arguments to be passed to \code{bs()}.}
}
\value{
An object of class "nof1.data" that is used to run the model using \code{\link{nof1.run}} is a list
containing
\item{Y}{Outcome}
\item{Treat}{Treatment}
\item{ncat}{Number of categories for ordinal response}
\item{nobs}{Total number of observations in a study}
\item{Treat.name}{Treatment name}
\item{response}{Type of outcome}
\item{Treat_\emph{Treat.name}}{Vector in the model matrix for \emph{Treat.name}}
\item{bs.trend}{Indicator for whether the model should adjust for trend using splines}
\item{corr.y}{Indicator for whether the correlation among measurements shoule be modeled}
\item{priors}{Priors that the code will be using. Default priors are used if prior was not specified}
\item{code}{Rjags model file code that is generated using information provided by the user. To view model file inside R, use \code{cat(nof1$code).}}
}
\description{
Make an N of 1 object containing data, priors, and a jags model file
}
\examples{
###Blocker data example
laughter
Y <- laughter$Y
Treat <- laughter$Treat
nof1 <- nof1.data(Y, Treat, ncat = 11, response = "ordinal")
str(nof1)
cat(nof1$code)
}
