% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nof1.summary.R
\name{summarize_nof1}
\alias{summarize_nof1}
\title{Function to present a summary of our results}
\usage{
summarize_nof1(result, alpha = 0.05)
}
\arguments{
\item{result}{A list which contains data file created
by \code{nof1.data} and the result file created by \code{nof1.run}}

\item{alpha}{The alpha value for the confidence interval. If no value is
entered will give the 95\% confidence interval.}
}
\value{
The function computes and returns a list of summary statistics of 
the raw data and the fitted model.
\item{raw.y.mean}{The raw mean of the outcome for each treatment}
\item{raw.y.median}{The raw median of the outcome for each treatment}
\item{post.coef.mean}{The posterior mean of the coefficient for each treatment}
\item{post.coef.median}{The posterior median of the coefficient for each treatment}
\item{post.y.mean}{The posterior mean of the outcome for each treatment}
\item{post.y.median}{The posterior median of the outcome for each treatment}
\item{post.coef.ci}{The credible interval of the coefficient for each treatment}
\item{post.y.ci}{The credible interval of the outcome for each treatment}
\item{comp.treat.post.coef}{The posterior quantiles of one coefficient minus the other 
when comparing two treatments}
\item{comp.treat.post.y}{The posterior quantiles of outcome minus the other when 
comparing two treatments}
\item{p.comp.coef.greater.0}{The posterior probability of one coefficient minus the other 
greater than 0}
}
\description{
A neat function to summarize the results.
}
