% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nof1.summary.R
\name{time_series_plot}
\alias{time_series_plot}
\title{time series plot across different interventions}
\usage{
time_series_plot(result, overlay.with.model = F, plot.by.treat = T,
  trial.start = NULL, trial.end = NULL, timestamp.format = NULL)
}
\arguments{
\item{trial.start}{Start time of the trial specified with \code{timestamp.format}.}

\item{trial.end}{End time of the trial specified with \code{timestamp.format}.}

\item{timestamp.format}{Format of the \code{trial.start} and \code{trial.end}.}

\item{nof1}{nof1 object created using nof1.data}
}
\description{
time series plot across different interventions
}
\examples{
Y <- laughter$Y
Treat <- laughter$Treat
nof1 <- nof1.data(Y, Treat, ncat = 11, baseline = "Usual Routine", response = "ordinal")
timestamp <- seq(as.Date('2015-01-01'),as.Date('2016-01-31'), length.out = length(Y))
time_series_plot(nof1, timestamp = timestamp, timestamp.format = "\%m-\%d-\%Y", Outcome.name = "Stress")
}
