% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nof1.summary.R
\name{trt_eff_plot}
\alias{trt_eff_plot}
\title{Errorbars for the credible interval of the treatment effect}
\usage{
trt_eff_plot(result.list, level = 0.95, ...)
}
\arguments{
\item{result.list}{A list of modeling results from \code{nof1.run}.}

\item{...}{parameters to pass to \code{geom_errorbar}.}
}
\description{
Creates errorbars for the credible interval of estimated treatment effect
}
